A lab and the final project
I did the lab a few days ago now and forgot to post it. It was just answering a couple questions about clustering. I've also done the final project for the course. It was just practicing making and training models that the course introduced me to. Using linear regression, KNN, trees, logistic regression, and an SVM.  The data used was weather data. Where the features columns was details about a certain day and the target column was predicting if there would be rain the next day or not. Most of the models performed moderate. I'm sure with proper hyper tuning the model could have gotten much better metrics. However, the assignment didn't ask to go further than just using basic modeling.