Linear Regression Network on Some Data
I've completed the capstone for the second course. It was nothing too complicated. Just importing some data, normalizing it, setting up a training loop to measure a mean of mean squared error for a linear regression model. The assignment didn't require me to reach a minimum, which is why I left it as it was. All I did was enough work to complete the parts of the assignment.