Course 3 Module 2 Labs Done
I've done the labs for the second module of course 3. The Labs consisted of just leaning about how to manipulate images if needed. I don't really think I personally will be using this. It's nice to know that python has libraries that allow devs to manipulate images easily if needed though.